[openstack] Added accumulated code analysis
This set of files reflect what we have done in the past few weeks while
we try to analyze code of keystone and horizon. One of the big change in
our approach is that we run the openstack modules in real hardware,
allowing us to actually analyze and debug the code at runtime.